+++ R/class_active.R
@@ -73,8 +73,8 @@ active_class <- R6::R6Class(
         remove(list = discard, envir = envir)
         out <- list(.tar_envir_5048826d = envir)
       }
-      out[[".tar_path_store_5048826d"]] <- path_store
       out[[".tar_options_5048826d"]] <- tar_options$export()
+      out[[".tar_path_store_5048826d"]] <- path_store
       out
     },
     unload_transient = function() {