+++ NEWS.md
@@ -1,4 +1,4 @@
-# targets 0.6.0.9000
+# targets 0.6.0.9001
 
 ## Bug fixes